Move DataCheckerPermission to skaianet package
